+++ Gems/Multiplayer/Code/Source/NetworkEntity/NetworkEntityUpdateMessage.cpp
@@ -128,7 +128,7 @@ namespace Multiplayer
         }
 
         // 2-byte size header + the actual blob payload itself
-        const uint32_t sizeOfBlob = (m_data != nullptr) ? sizeof(PropertyIndex) + m_data->GetSize() : 0;
+        const uint32_t sizeOfBlob = static_cast<uint32_t>((m_data != nullptr) ? sizeof(PropertyIndex) + m_data->GetSize() : 0);
 
         if (m_hasValidPrefabId)
         {
